add resize support

+++ hack/e2e.sh
@@ -17,7 +17,9 @@ function cleanup(){
     echo "Cleaning Namespace"
     kubectl delete secret qcsecret -n $TEST_NS
     kubectl delete ns $TEST_NS > /dev/null
-    rm -f $secret_file
+    if [ x$ACCESS_KEY_ID != "x" ]; then
+        rm -f $secret_file
+    fi
     if [ $SKIP_BUILD == "no" ]; then
         docker  rmi $IMG
     fi
@@ -70,7 +72,7 @@ fi
 sed -e 's@image: .*@image: '"${IMG}"'@' -e 's/kube-system/'"$TEST_NS"'/g' deploy/kube-cloud-controller-manager.yaml > $DEST
 
 kubectl create ns $TEST_NS
-if [ $ACCESS_KEY_ID != "" ]; then
+if [ x$ACCESS_KEY_ID != "x" ]; then
     printf "qy_access_key_id: '%s'\nqy_secret_access_key: '%s'\nzone: 'ap2a'\n" $ACCESS_KEY_ID $SECRET_ACCESS_KEY > /tmp/config.yaml
     secret_file=/tmp/config.yaml
 fi
@@ -78,5 +80,6 @@ fi
 kubectl create secret generic qcsecret --from-file=$secret_file -n $TEST_NS
 kubectl apply -f $DEST
 export TEST_NS
+export secret_file
 
 go test -v -mod=vendor ./test/pkg/e2e/
